requirements.txt
matplotlib==3.6.2
numpy==1.24.4
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
